Export timeunit.format() for VgTooltip
